Hard-code LIB_VERSION for now because it breaks inside packaged Atom
We delete the package.json of all bundled packages during our slug
compilation step. I don	’t feel comfortable changing that on stable,
but we’ll need exception reporting to work in the meantime. Once we
figure out if we can retain the package.json files, we can restore the
original approach.

Signed-off-by: Max Brunsfeld <maxbrunsfeld@github.com>